feat(#768): rename AnimationsGalleryView to EffectsGalleryView
git mv + rewrite: every Animation* identifier renamed to Effect*
(EffectsGalleryModel, EffectsGalleryView, EffectDetailView,
EffectDemoWebView), navigation title and empty-state copy updated to
"Effects", and EffectCategory.displayName extended with the four new
categories from Task 1 (Canvas Backgrounds, Cursor-Reactive,
Scroll-Driven, Generative Art).

Sidebar grouping: used a flat Section per category (not nested
Section-in-Section) with a plain Text group-header row above each
group's first category, per the brief's fallback option. Chose this
over nesting because whether a two-level Section hierarchy renders
both header levels in a macOS sidebar List can only be confirmed
visually in a running app, which wasn't available while authoring
this change; the flat-Section-plus-header-row construct is the one
known to render correctly.

Adds the Apply-to-page flow (Task 10's EffectPlacementController):
EffectsGalleryView now takes an injected controller plus
enterOverlayMode/exitOverlayMode closures (threaded through the same
way demoURL already was), EffectDetailView gets an "Apply to
Page..." button next to Copy Snippet shown only when
entry.placement != nil, and a placement HUD (ZStack overlay) shows
"Click where to place this effect" + Cancel during
.picking/.applying, with a transient success/failure banner for
.succeeded/.failed (controller.cancel() only rewinds .picking, so
the banner auto-hides on a local timer rather than resetting
controller state). Esc is wired via .onExitCommand.

SiteWindow.swift still calls the old AnimationsGalleryView() with no
arguments and is now expected to fail to compile - wiring the real
controller and overlay-mode closures into SiteWindow/SiteWindowModel
is Task 12's job, not touched here. Confirmed the only compiler
errors anywhere in the package are both in SiteWindow.swift and both
trace to that one missing symbol: a "cannot find 'AnimationsGalleryView'
in scope" at its call site, plus a knock-on "unable to type-check
this expression in reasonable time" at an unrelated line (805) that
disappears once the missing symbol resolves - verified by temporarily
shimming the old name back in and rebuilding clean.
